Simple uility that outputs data to stdout and peer ID to stderr
